Refactors shared Ethereum client features into new class EthereumClient
